chore: move RegexIconButton
